+++ catalogue-microservice/src/main/java/catalogue/microsservice/cataloguemicroservice/repository/KatalogRepository.java
@@ -1,13 +1,11 @@
 package catalogue.microsservice.cataloguemicroservice.repository;
 
 import catalogue.microsservice.cataloguemicroservice.model.Katalog;
-import catalogue.microsservice.cataloguemicroservice.model.Korisnik;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface KatalogRepository extends JpaRepository<Katalog, Integer> {
+public interface KatalogRepository extends JpaRepository<Katalog, Long> {
 }
