+++ catalogue-microservice/src/main/java/catalogue/microsservice/cataloguemicroservice/repository/KorisnikRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface KorisnikRepository extends JpaRepository<Korisnik, Integer> {
+public interface KorisnikRepository extends JpaRepository<Korisnik, Long> {
 }
